Move postgres-specific state to own module
This change moves the various SqliteBackend-specific items to a postgres
module, `transact::state::merkle::sql::postgres`.
