+++ libtransact/src/state/merkle/sql/mod.rs
@@ -53,6 +53,8 @@ mod error;
 pub mod migration;
 mod models;
 mod operations;
+#[cfg(feature = "postgres")]
+mod postgres;
 mod schema;
 #[cfg(feature = "sqlite")]
 mod sqlite;
@@ -60,27 +62,14 @@ mod sqlite;
 use std::cmp::Reverse;
 use std::collections::{HashMap, HashSet};
 
-use diesel::Connection as _;
-
-use crate::error::{InternalError, InvalidStateError};
-use crate::state::error::{StatePruneError, StateReadError, StateWriteError};
-use crate::state::merkle::{MerkleRadixLeafReadError, MerkleRadixLeafReader};
-use crate::state::{Prune, Read, StateChange, Write};
+use crate::error::InternalError;
+use crate::state::error::StateWriteError;
+use crate::state::StateChange;
 
 use super::node::Node;
 
-use backend::{Backend, Connection};
+use backend::Backend;
 pub use error::SqlMerkleStateBuildError;
-use operations::get_leaves::MerkleRadixGetLeavesOperation as _;
-use operations::get_or_create_tree::MerkleRadixGetOrCreateTreeOperation as _;
-use operations::get_path::MerkleRadixGetPathOperation as _;
-use operations::get_tree_by_name::MerkleRadixGetTreeByNameOperation as _;
-use operations::has_root::MerkleRadixHasRootOperation as _;
-use operations::insert_nodes::MerkleRadixInsertNodesOperation as _;
-use operations::list_leaves::MerkleRadixListLeavesOperation as _;
-use operations::prune_entries::MerkleRadixPruneEntriesOperation as _;
-use operations::update_change_log::MerkleRadixUpdateUpdateChangeLogOperation as _;
-use operations::MerkleRadixOperations;
 
 const TOKEN_SIZE: usize = 2;
 
@@ -121,43 +110,6 @@ impl<B: Backend + Clone> SqlMerkleStateBuilder<B> {
     }
 }
 
-#[cfg(feature = "postgres")]
-impl SqlMerkleStateBuilder<backend::PostgresBackend> {
-    /// Construct the final SqlMerkleState instance
-    ///
-    /// # Errors
-    ///
-    /// An error may be returned under the following circumstances:
-    /// * If a Backend has not been provided
-    /// * If a tree name has not been provided
-    /// * If an internal error occurs while trying to create or lookup the tree
-    pub fn build(
-        self,
-    ) -> Result<SqlMerkleState<backend::PostgresBackend>, SqlMerkleStateBuildError> {
-        let backend = self
-            .backend
-            .ok_or_else(|| InvalidStateError::with_message("must provide a backend".into()))?;
-
-        let tree_name = self
-            .tree_name
-            .ok_or_else(|| InvalidStateError::with_message("must provide a tree name".into()))?;
-
-        let conn = backend.connection()?;
-        let operations = MerkleRadixOperations::new(conn.as_inner());
-
-        let (initial_state_root_hash, _) = encode_and_hash(Node::default())?;
-        let tree_id: i64 = if self.create_tree {
-            operations.get_or_create_tree(&tree_name, &hex::encode(&initial_state_root_hash))?
-        } else {
-            operations.get_tree_id_by_name(&tree_name)?.ok_or_else(|| {
-                InvalidStateError::with_message("must provide the name of an existing tree".into())
-            })?
-        };
-
-        Ok(SqlMerkleState { backend, tree_id })
-    }
-}
-
 /// SqlMerkleState provides a merkle-radix implementation over a SQL database.
 ///
 /// Databases are implemented using a Backend to provide connections. Note that the database must
@@ -176,124 +128,6 @@ impl<B: Backend + Clone> SqlMerkleState<B> {
         let (hash, _) = encode_and_hash(Node::default())?;
 
         Ok(hex::encode(&hash))
-    }
-}
-
-#[cfg(feature = "postgres")]
-impl Write for SqlMerkleState<backend::PostgresBackend> {
-    type StateId = String;
-    type Key = String;
-    type Value = Vec<u8>;
-
-    fn commit(
-        &self,
-        state_id: &Self::StateId,
-        state_changes: &[StateChange],
-    ) -> Result<Self::StateId, StateWriteError> {
-        let overlay =
-            MerkleRadixOverlay::new(self.tree_id, &*state_id, SqlOverlay::new(&self.backend));
-
-        let (next_state_id, tree_update) = overlay
-            .generate_updates(state_changes)
-            .map_err(|e| StateWriteError::StorageError(Box::new(e)))?;
-
-        overlay
-            .write_updates(&next_state_id, tree_update)
-            .map_err(|e| StateWriteError::StorageError(Box::new(e)))?;
-
-        Ok(next_state_id)
-    }
-
-    fn compute_state_id(
-        &self,
-        state_id: &Self::StateId,
-        state_changes: &[StateChange],
-    ) -> Result<Self::StateId, StateWriteError> {
-        let overlay =
-            MerkleRadixOverlay::new(self.tree_id, &*state_id, SqlOverlay::new(&self.backend));
-
-        let (next_state_id, _) = overlay
-            .generate_updates(state_changes)
-            .map_err(|e| StateWriteError::StorageError(Box::new(e)))?;
-
-        Ok(next_state_id)
-    }
-}
-
-#[cfg(feature = "postgres")]
-impl Prune for SqlMerkleState<backend::PostgresBackend> {
-    type StateId = String;
-    type Key = String;
-    type Value = Vec<u8>;
-
-    fn prune(&self, state_ids: Vec<Self::StateId>) -> Result<Vec<Self::Key>, StatePruneError> {
-        let overlay = MerkleRadixPruner::new(self.tree_id, SqlOverlay::new(&self.backend));
-
-        overlay
-            .prune(&state_ids)
-            .map_err(|e| StatePruneError::StorageError(Box::new(e)))
-    }
-}
-
-#[cfg(feature = "postgres")]
-impl Read for SqlMerkleState<backend::PostgresBackend> {
-    type StateId = String;
-    type Key = String;
-    type Value = Vec<u8>;
-
-    fn get(
-        &self,
-        state_id: &Self::StateId,
-        keys: &[Self::Key],
-    ) -> Result<HashMap<Self::Key, Self::Value>, StateReadError> {
-        let overlay =
-            MerkleRadixOverlay::new(self.tree_id, &*state_id, SqlOverlay::new(&self.backend));
-
-        if !overlay
-            .has_root()
-            .map_err(|e| StateReadError::StorageError(Box::new(e)))?
-        {
-            return Err(StateReadError::InvalidStateId(state_id.into()));
-        }
-
-        overlay
-            .get_entries(keys)
-            .map_err(|e| StateReadError::StorageError(Box::new(e)))
-    }
-
-    fn clone_box(
-        &self,
-    ) -> Box<dyn Read<StateId = Self::StateId, Key = Self::Key, Value = Self::Value>> {
-        Box::new(self.clone())
-    }
-}
-
-type IterResult<T> = Result<T, MerkleRadixLeafReadError>;
-type LeafIter<T> = Box<dyn Iterator<Item = IterResult<T>>>;
-
-#[cfg(feature = "postgres")]
-impl MerkleRadixLeafReader for SqlMerkleState<backend::PostgresBackend> {
-    /// Returns an iterator over the leaves of a merkle radix tree.
-    /// By providing an optional address prefix, the caller can limit the iteration
-    /// over the leaves in a specific subtree.
-    fn leaves(
-        &self,
-        state_id: &Self::StateId,
-        subtree: Option<&str>,
-    ) -> IterResult<LeafIter<(Self::Key, Self::Value)>> {
-        let conn = self.backend.connection()?;
-
-        if &self.initial_state_root_hash()? == state_id {
-            return Ok(Box::new(std::iter::empty()));
-        }
-
-        let leaves = MerkleRadixOperations::new(conn.as_inner()).list_leaves(
-            self.tree_id,
-            state_id,
-            subtree,
-        )?;
-
-        Ok(Box::new(leaves.into_iter().map(Ok)))
     }
 }
 
@@ -560,96 +394,6 @@ struct SqlOverlay<'b, B: Backend> {
 impl<'b, B: Backend> SqlOverlay<'b, B> {
     fn new(backend: &'b B) -> Self {
         Self { backend }
-    }
-}
-
-#[cfg(feature = "postgres")]
-impl<'b> OverlayReader for SqlOverlay<'b, backend::PostgresBackend> {
-    fn has_root(&self, tree_id: i64, state_root_hash: &str) -> Result<bool, InternalError> {
-        let conn = self.backend.connection()?;
-
-        let operations = MerkleRadixOperations::new(conn.as_inner());
-        operations.has_root(tree_id, state_root_hash)
-    }
-
-    fn get_path(
-        &self,
-        tree_id: i64,
-        state_root_hash: &str,
-        address: &str,
-    ) -> Result<Vec<(String, Node)>, InternalError> {
-        let conn = self.backend.connection()?;
-
-        let operations = MerkleRadixOperations::new(conn.as_inner());
-        operations.get_path(tree_id, state_root_hash, address)
-    }
-
-    fn get_entries(
-        &self,
-        tree_id: i64,
-        state_root_hash: &str,
-        keys: Vec<&str>,
-    ) -> Result<Vec<(String, Vec<u8>)>, InternalError> {
-        let conn = self.backend.connection()?;
-
-        let operations = MerkleRadixOperations::new(conn.as_inner());
-        operations.get_leaves(tree_id, state_root_hash, keys)
-    }
-}
-
-#[cfg(feature = "postgres")]
-impl<'b> OverlayWriter for SqlOverlay<'b, backend::PostgresBackend> {
-    fn write_changes(
-        &self,
-        tree_id: i64,
-        state_root_hash: &str,
-        parent_state_root_hash: &str,
-        tree_update: TreeUpdate,
-    ) -> Result<(), InternalError> {
-        let conn = self.backend.connection()?;
-
-        conn.as_inner().transaction(|| {
-            let operations = MerkleRadixOperations::new(conn.as_inner());
-
-            let TreeUpdate {
-                node_changes,
-                deletions,
-            } = tree_update;
-
-            let insertable_changes = node_changes
-                .into_iter()
-                .map(
-                    |(hash, node, address)| operations::insert_nodes::InsertableNode {
-                        hash,
-                        node,
-                        address,
-                    },
-                )
-                .collect::<Vec<_>>();
-
-            operations.insert_nodes(tree_id, &insertable_changes)?;
-
-            let additions = insertable_changes
-                .iter()
-                .map(|insertable| insertable.hash.as_ref())
-                .collect::<Vec<_>>();
-            let deletions = deletions.iter().map(|s| s.as_ref()).collect::<Vec<_>>();
-            operations.update_change_log(
-                tree_id,
-                state_root_hash,
-                parent_state_root_hash,
-                &additions,
-                &deletions,
-            )?;
-
-            Ok(())
-        })
-    }
-
-    fn prune(&self, tree_id: i64, state_root: &str) -> Result<Vec<String>, InternalError> {
-        let conn = self.backend.connection()?;
-        let operations = MerkleRadixOperations::new(conn.as_inner());
-        operations.prune_entries(tree_id, state_root)
     }
 }
